Issue #2790107 by larowlan: Workbench Access for all content entity types - menu access working
